% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstudentSD.R
\name{dtstudentSD}
\alias{ddtstudentSD}
\alias{dtstudentSD}
\title{Discrete Truncated Student-t Distribution, Standard Deviation Parameterization}
\usage{
ddtstudentSD(x, nuprime, mu, sigma, width = 1, lower = 0, upper = 100,
  log = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{nuprime}{\code{degrees of freedom - 2} of underlying Student-t distribution. \code{nuprime > 0},
and as \code{nu} goes to \code{Inf}, the underlying distribution becomes Normal.}

\item{mu}{mode / mean of underlying Student-t distribution.}

\item{sigma}{standard deviation of underlying Student-t distribution.}

\item{width}{width of the intervals used for discretization.}

\item{lower}{lower truncation point.}

\item{upper}{upper truncation point.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \emph{P[X ≤ x]} otherwise,
\emph{P[X > x]}.}
}
\description{
Discrete truncated Student-t distribution parameterized by standard deviation.
}
\author{
Matthew Kay
}

